coap: Split impl from helper types
